Basic concept gameplay
Fixed the issue experienced with collisions not working. When physics
are enabled, normal features does not work the same and require more
research.

This push contains a basic concept gameplay.
Create a water bubble.
Water bubble floats to the right.
Score when bubble pops.
Water drops onto fire and destroy the fire.